# api/docs/swagger.yaml
definitions:
...
  auth_service.ClientPlatform:
    properties:
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  auth_service.ClientType:
    properties:
      confirm_by:
        type: integer
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      self_recover:
        type: boolean
      self_register:
        type: boolean
      tables:
        items:
          $ref: '#/definitions/auth_service.Table'
        type: array
    type: object
...
  auth_service.HasAccessRequest:
    properties:
      access_token:
        type: string
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      project_id:
        type: string
    type: object
...
  auth_service.LoginRequest:
    properties:
      password:
        type: string
      tables:
        items:
          $ref: '#/definitions/auth_service.Object'
        type: array
      username:
        type: string
    type: object
...
  auth_service.LogoutRequest:
    properties:
      access_token:
        type: string
    type: object
  auth_service.Object:
    properties:
      object_id:
        type: string
      table_slug:
        type: string
    type: object
...
  auth_service.RecordPermission:
    properties:
      client_type_id:
        type: string
      delete:
        type: string
      id:
        type: string
      read:
        type: string
      role_id:
        type: string
      table_slug:
        type: string
      update:
        type: string
      write:
        type: string
    type: object
  auth_service.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  auth_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
...
  auth_service.Session:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      data:
        type: string
      expires_at:
        type: string
      id:
        type: string
      integration_id:
        type: string
      ip:
        type: string
      is_changed:
        type: boolean
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  auth_service.Table:
    properties:
      data:
        type: string
      icon:
        type: string
      label:
        type: string
      slug:
        type: string
      view_label:
        type: string
      view_slug:
        type: string
    type: object
  auth_service.Token:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      refresh_in_seconds:
        type: integer
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
...
  auth_service.User:
    properties:
      active:
        type: integer
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: string
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      photo_url:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
    type: object
...
  auth_service.V2LoginRequest:
    properties:
      client_type:
        type: string
      login_strategy:
        type: string
      password:
        type: string
      tables:
        items:
          $ref: '#/definitions/auth_service.Object'
        type: array
      username:
        type: string
    type: object
  auth_service.V2LoginResponse:
    properties:
      client_platform:
        $ref: '#/definitions/auth_service.ClientPlatform'
        type: object
      client_type:
        $ref: '#/definitions/auth_service.ClientType'
        type: object
      permissions:
        items:
          $ref: '#/definitions/auth_service.RecordPermission'
        type: array
      role:
        $ref: '#/definitions/auth_service.Role'
        type: object
      sessions:
        items:
          $ref: '#/definitions/auth_service.Session'
        type: array
      token:
        $ref: '#/definitions/auth_service.Token'
        type: object
      user_found:
        type: boolean
      user_id:
        type: string
    type: object
  auth_service.V2RefreshTokenResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/auth_service.RecordPermission'
        type: array
      token:
        $ref: '#/definitions/auth_service.Token'
        type: object
    type: object
  config.Config:
    properties:
      authGRPCPort:
        type: string
      authServiceHost:
        type: string
      defaultLimit:
        type: string
      defaultOffset:
        type: string
      environment:
        description: debug, test, release
        type: string
      httpport:
        type: string
      httpscheme:
        type: string
      objectBuilderGRPCPort:
        type: string
      objectBuilderServiceHost:
        type: string
      passcodeLength:
        type: integer
      passcodePool:
        type: string
      postgresDatabase:
        type: string
      postgresHost:
        type: string
      postgresMaxConnections:
        type: integer
      postgresPassword:
        type: string
      postgresPort:
        type: integer
      postgresUser:
        type: string
      secretKey:
        type: string
      serviceName:
        type: string
      smsGRPCPort:
        type: string
      smsServiceHost:
        type: string
      version:
        type: string
    type: object
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
...
  models.RegisterOtp:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
  models.SendCodeResponse:
    properties:
      data:
        $ref: '#/definitions/object_builder_service.V2LoginResponse'
        type: object
      sms_id:
        type: string
    type: object
  models.Sms:
    properties:
      client_type:
        type: string
      recipient:
        type: string
      text:
        type: string
    type: object
...
  models.Verify:
    properties:
      data:
        $ref: '#/definitions/object_builder_service.V2LoginResponse'
        type: object
      tables:
        items:
          $ref: '#/definitions/auth_service.Object'
        type: array
    type: object
...
  object_builder_service.ClientPlatform:
    properties:
      guid:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
  object_builder_service.ClientType:
    properties:
      confirm_by:
        type: integer
      guid:
        type: string
      name:
        type: string
      project_id:
        type: string
      self_recover:
        type: boolean
      self_register:
        type: boolean
      tables:
        items:
          $ref: '#/definitions/object_builder_service.TableClientType'
        type: array
    type: object
  object_builder_service.RecordPermission:
    properties:
      client_type_id:
        type: string
      delete:
        type: string
      guid:
        type: string
      read:
        type: string
      role_id:
        type: string
      table_slug:
        type: string
      update:
        type: string
      write:
        type: string
    type: object
  object_builder_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      guid:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
  object_builder_service.TableClientType:
    properties:
      data:
        type: string
      icon:
        type: string
      label:
        type: string
      slug:
        type: string
      view_label:
        type: string
      view_slug:
        type: string
    type: object
  object_builder_service.V2LoginResponse:
    properties:
      client_platform:
        $ref: '#/definitions/object_builder_service.ClientPlatform'
        type: object
      client_type:
        $ref: '#/definitions/object_builder_service.ClientType'
        type: object
      permissions:
        items:
          $ref: '#/definitions/object_builder_service.RecordPermission'
        type: array
      role:
        $ref: '#/definitions/object_builder_service.Role'
        type: object
      user_found:
        type: boolean
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
...
  /config:
    get:
      consumes:
      - application/json
      description: show service config data when the service environment set to debug
        mode
      operationId: get_config
      produces:
      - application/json
      responses:
        "200":
          description: Response data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/config.Config'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Response'
      summary: get config data on the debug mode
  /has-access:
    post:
      consumes:
      - application/json
      description: Has Access
      operationId: has_access
      parameters:
      - description: HasAccessRequestBody
        in: body
        name: has-access
        required: true
        schema:
          $ref: '#/definitions/auth_service.HasAccessRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Has Access
      tags:
      - Session
...
  /login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth_service.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - Session
  /logout:
    delete:
      consumes:
      - application/json
      description: Logout User
      operationId: logout
      parameters:
      - description: LogoutRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_service.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Logout User
      tags:
      - Session
...
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh Token
      operationId: refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Token
      tags:
      - Session
...
  /register-otp/{table_slug}:
    post:
      consumes:
      - application/json
      description: RegisterOtp
      operationId: registerOtp
      parameters:
      - description: register_body
        in: body
        name: registerBody
        required: true
        schema:
          $ref: '#/definitions/models.RegisterOtp'
      - description: table_slug
        in: path
        name: table_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.V2LoginResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: RegisterOtp
      tags:
      - register
...
  /send-code:
    post:
      consumes:
      - application/json
      description: SendCode
      operationId: sendCode
      parameters:
      - description: SendCode
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.Sms'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SendCodeResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: SendCode
      tags:
      - register
...
  /v2/login:
    post:
      consumes:
      - application/json
      description: V2Login
      operationId: V2login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth_service.V2LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: V2Login
      tags:
      - V2_Session
...
  /v2/refresh:
    put:
      consumes:
      - application/json
      description: V2Refresh Token
      operationId: v2refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.V2RefreshTokenResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: V2Refresh Token
      tags:
      - V2_Session
...
  /verify/{sms_id}/{otp}:
    post:
      consumes:
      - application/json
      description: Verify
      operationId: verify
      parameters:
      - description: sms_id
        in: path
        name: sms_id
        required: true
        type: string
      - description: otp
        in: path
        name: otp
        required: true
        type: string
      - description: verify_body
        in: body
        name: verifyBody
        required: true
        schema:
          $ref: '#/definitions/models.Verify'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.V2LoginResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Verify
      tags:
      - register
swagger: "2.0"