# api/handlers/register.go
package handlers

import (
	"context"
	"errors"
	"ucode/ucode_go_auth_service/api/http"
	"ucode/ucode_go_auth_service/api/models"
	pb "ucode/ucode_go_auth_service/genproto/auth_service"
	pbObject "ucode/ucode_go_auth_service/genproto/object_builder_service"
	pbSms "ucode/ucode_go_auth_service/genproto/sms_service"
	"ucode/ucode_go_auth_service/pkg/helper"
	"ucode/ucode_go_auth_service/pkg/util"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

// SendCode godoc
// @ID sendCode
// @Router /send-code [POST]
// @Summary SendCode
// @Description SendCode
// @Tags register
// @Accept json
// @Produce json
// @Param login body models.Sms true "SendCode"
// @Success 201 {object} http.Response{data=models.SendCodeResponse} "User data"
// @Response 400 {object} http.Response{data=string} "Bad Request"
// @Failure 500 {object} http.Response{data=string} "Server Error"
func (h *Handler) SendCode(c *gin.Context) {

	var request models.Sms

	err := c.ShouldBindJSON(&request)
	if err != nil {
		h.handleResponse(c, http.BadRequest, err.Error())
		return
	}
	id, err := uuid.NewRandom()
	if err != nil {
		h.handleResponse(c, http.InternalServerError, err.Error())
		return
	}
	valid := util.IsValidPhone(request.Recipient)
	if !valid {
		h.handleResponse(c, http.BadRequest, "Неверный номер телефона, он должен содержать двенадцать цифр и +")
		return
	}

	expire := time.Now().Add(time.Minute * 5)

	code, err := util.GenerateCode(4)
	if err != nil {
		h.handleResponse(c, http.InternalServerError, err.Error())
		return
	}

	phone := helper.ConverPhoneNumberToMongoPhoneFormat(request.Recipient)

	respObject, err := h.services.LoginService().LoginWithOtp(c.Request.Context(), &pbObject.PhoneOtpRequst{
		PhoneNumber: phone,
		ClientType:  request.ClientType,
	})
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	if (respObject == nil || !respObject.UserFound) && request.ClientType != "PATIENT" {
		err := errors.New("Пользователь не найдено")
		h.handleResponse(c, http.NotFound, err.Error())
		return
	}
	resp, err := h.services.SmsService().Send(
		c.Request.Context(),
		&pbSms.Sms{
			Id:        id.String(),
			Text:      request.Text,
			Otp:       code,
			Recipient: request.Recipient,
			ExpiresAt: expire.String()[:19],
		},
	)
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}
	res := models.SendCodeResponse{
		SmsId: resp.SmsId,
		Data:  respObject,
	}

	h.handleResponse(c, http.Created, res)

}

// Verify godoc
// @ID verify
// @Router /verify/{sms_id}/{otp} [POST]
// @Summary Verify
// @Description Verify
// @Tags register
// @Accept json
// @Produce json
// @Param sms_id path string true "sms_id"
// @Param otp path string true "otp"
// @Param verifyBody body models.Verify true "verify_body"
// @Success 201 {object} http.Response{data=auth_service.V2LoginResponse} "User data"
// @Response 400 {object} http.Response{data=string} "Bad Request"
// @Failure 500 {object} http.Response{data=string} "Server Error"
func (h *Handler) Verify(c *gin.Context) {
	var body models.Verify

	_, err := h.services.SmsService().ConfirmOtp(
		c.Request.Context(),
		&pbSms.ConfirmOtpRequest{
			SmsId: c.Param("sms_id"),
			Otp:   c.Param("otp"),
		},
	)
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	err = c.ShouldBindJSON(&body)
	if err != nil {
		h.handleResponse(c, http.BadRequest, err.Error())
		return
	}
	if !body.Data.UserFound {
		h.handleResponse(c, http.OK, "User verified but not found")
		return
	}
	convertedToAuthPb := helper.ConvertPbToAnotherPb(body.Data)
	res, err := h.services.SessionService().SessionAndTokenGenerator(context.Background(), &pb.SessionAndTokenRequest{
		LoginData: convertedToAuthPb,
		Tables:    body.Tables,
	})
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	h.handleResponse(c, http.Created, res)
}

// RegisterOtp godoc
// @ID registerOtp
// @Router /register-otp/{table_slug} [POST]
// @Summary RegisterOtp
// @Description RegisterOtp
// @Tags register
// @Accept json
// @Produce json
// @Param registerBody body models.RegisterOtp true "register_body"
// @Param table_slug path string true "table_slug"
// @Success 201 {object} http.Response{data=auth_service.V2LoginResponse} "User data"
// @Response 400 {object} http.Response{data=string} "Bad Request"
// @Failure 500 {object} http.Response{data=string} "Server Error"
func (h *Handler) RegisterOtp(c *gin.Context) {
	var body models.RegisterOtp

	err := c.ShouldBindJSON(&body)
	if err != nil {
		h.handleResponse(c, http.BadRequest, err.Error())
		return
	}

	structData, err := helper.ConvertMapToStruct(body.Data)

	if err != nil {
		h.handleResponse(c, http.InvalidArgument, err.Error())
		return
	}
	_, err = h.services.ObjectBuilderService().Create(
		context.Background(),
		&pbObject.CommonMessage{
			TableSlug: c.Param("table_slug"),
			Data:      structData,
		},
	)
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	resp, err := h.services.LoginService().LoginWithOtp(context.Background(), &pbObject.PhoneOtpRequst{
		PhoneNumber: body.Data["phone_number"].(string),
		ClientType:  "PATIENT",
	})
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	convertedToAuthPb := helper.ConvertPbToAnotherPb(resp)
	res, err := h.services.SessionService().SessionAndTokenGenerator(context.Background(), &pb.SessionAndTokenRequest{
		LoginData: convertedToAuthPb,
		Tables:    []*pb.Object{},
	})
	if err != nil {
		h.handleResponse(c, http.GRPCError, err.Error())
		return
	}

	h.handleResponse(c, http.Created, res)
}